Revert "fix: don't assume vertices are uint32"
This reverts commit f920de3c6d3feb3c09949319572354a2110b5256.

+++ kimimaro/trace.py
@@ -161,7 +161,7 @@ def trace(
     [ PrecomputedSkeleton.from_path(path) for path in paths ]
   ).consolidate()
 
-  verts = skel.vertices.flatten()
+  verts = skel.vertices.flatten().astype(np.uint32)
   skel.radii = DBF[verts[::3], verts[1::3], verts[2::3]]
 
   return skel
